Move fetch into tool submodules
