--- crates/volta-core/src/fetch/mod.rs
@@ -1,23 +1,0 @@
-//! Provides methods for fetching tool distributions into the local inventory.
-
-use crate::error::ErrorDetails;
-use crate::tool;
-
-mod node;
-mod package;
-mod serial;
-mod yarn;
-
-pub use node::fetch as node;
-pub use node::load_default_npm_version;
-pub use package::fetch as package;
-pub use package::{BinConfig, BinLoader};
-pub use yarn::fetch as yarn;
-
-fn download_tool_error(
-    tool: tool::Spec,
-    from_url: impl AsRef<str>,
-) -> impl FnOnce(&failure::Error) -> ErrorDetails {
-    let from_url = from_url.as_ref().to_string();
-    |_| ErrorDetails::DownloadToolNetworkError { tool, from_url }
-}
