+++ crates/volta-core/src/lib.rs
@@ -8,7 +8,6 @@ mod command;
 pub mod env;
 pub mod error;
 mod event;
-pub(crate) mod fetch;
 pub(crate) mod fs;
 mod hook;
 pub mod inventory;
